+++ merged-packages/eth-json-rpc-middleware/src/types.test-d.ts
@@ -1,9 +1,10 @@
-import type { AbstractRpcService as NetworkControllerAbstractRpcService } from '@metamask/network-controller';
+import type { AbstractRpcService } from '@metamask/network-controller';
 import { expectAssignable } from 'tsd';
 
-import type { AbstractRpcService as LocalAbstractRpcService } from './types';
+import type { AbstractRpcServiceLike } from './types';
 
-// Confirm that the AbstractRpcService in this repo is compatible with the same
-// one in `@metamask/network-controller` (from where it was copied)
-declare const rpcService: LocalAbstractRpcService;
-expectAssignable<NetworkControllerAbstractRpcService>(rpcService);
+// Confirm that the `AbstractRpcServiceLike` type in this repo is compatible
+// with the type of the `request` method from `AbstractRpcService` in
+// `@metamask/network-controller` (from where it was copied).
+declare const networkControllerRpcService: AbstractRpcService;
+expectAssignable<AbstractRpcServiceLike>(networkControllerRpcService);